refactor empty screen
